# docs/openapi.yaml
openapi: 3.0.3
info:
  title: QMemory Node.js Utility Library API
  description: |
    Production-ready Node.js utility library providing comprehensive database operations,
    caching, circuit breaker patterns, and scalability enhancements.

    Features:
    - Database connection pooling with circuit breaker protection
    - Distributed tracing with correlation IDs
    - Rate limiting for production usage
    - Performance monitoring and metrics
    - Background I/O optimization
    - Request/response caching
    - Enhanced error handling and logging

    Architecture:
    - RESTful API design
    - Stateless for horizontal scaling
    - Type-safe TypeScript implementation
    - Production-grade security measures
  version: 2.1.0
  contact:
    name: QMemory Development Team
    email: support@qmemory.dev
  license: MIT
  servers:
    - url: https://api.qmemory.dev/v2
      description: Production server
    - url: https://staging-api.qmemory.dev/v2
      description: Staging server
    - url: http://localhost:5000/v2
      description: Development server
paths:
...
  /users:
    get:
      tags:
        - User Management
      summary: List users with pagination
      description: |
        Retrieves paginated list of users with optional filtering and sorting.
        Supports efficient pagination for large datasets.
      operationId: listUsers
      parameters:
        - name: page
          in: query
          description: Page number for pagination (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page (max 50)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
        - name: X-Request-ID
          in: header
          description: Correlation ID for distributed tracing
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-rate-limit: 1000 requests/minute

    post:
      tags:
        - User Management
      summary: Create new user
      description: |
        Creates a new user with validation and duplicate checking.
        Supports rate limiting and distributed tracing.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: User with this username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-rate-limit: 100 requests/minute
  /users/{userId}:
    get:
      tags:
        - User Management
      summary: Get user by ID
      description: |
        Retrieves a specific user by their numeric ID with validation.
        Includes detailed response headers for tracing.
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          description: Numeric user ID
          schema:
            type: integer
            minimum: 1
        - name: X-Request-ID
          in: header
          description: Correlation ID for distributed tracing
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      x-rate-limit: 1000 requests/minute

    put:
      tags:
        - User Management
      summary: Update user by ID
      description: |
        Updates user information with partial updates and validation.
        Supports partial field updates and maintains audit trail.
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          description: Numeric user ID
          schema:
            type: integer
            minimum: 1
        - name: X-Request-ID
          in: header
          description: Correlation ID for distributed tracing
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Username conflict if trying to update to existing username
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-rate-limit: 1000 requests/minute

    delete:
      tags:
        - User Management
      summary: Delete user by ID
      description: |
        Deletes a user with validation and cleanup of related resources.
        Includes soft delete capabilities for data recovery.
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          description: Numeric user ID
          schema:
            type: integer
            minimum: 1
        - name: X-Request-ID
          in: header
          description: Correlation ID for distributed tracing
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-rate-limit: 1000 requests/minute
  /users/by-username/{username}:
    get:
      tags:
        - User Management
      summary: Get user by username
      description: |
        Retrieves user information using username lookup with validation.
        Optimized for username-based searches.
      operationId: getUserByUsername
      parameters:
        - name: username
          in: path
          required: true
          description: User username (case-sensitive)
          schema:
            type: string
            minLength: 1
            maxLength: 50
            pattern: '^[a-zA-Z0-9_-]+$'
        - name: X-Request-ID
          in: header
          description: Correlation ID for distributed tracing
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
      x-rate-limit: 1000 requests/minute
...
components:
  schemas:
    User:
      type: object
      required:
        - id
        - username
      properties:
        id:
          type: integer
          description: Unique user identifier
          example: 12345
        username:
          type: string
          description: Unique username (1-50 chars, alphanumeric + underscore + hyphen)
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]+$'
          example: johndoe
        displayName:
          type: string
          description: Display name (1-100 chars)
          nullable: true
          minLength: 1
          maxLength: 100
          example: John Doe
        githubId:
          type: string
          description: GitHub username (optional)
          nullable: true
          example: johndoe
        avatar:
          type: string
          description: Avatar URL (optional)
          nullable: true
          example: https://example.com/avatar.jpg
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          readOnly: true
    CreateUserRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          description: Username (1-50 chars)
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]+$'
          example: newuser123
        displayName:
          type: string
          description: Display name (optional)
          nullable: true
          maxLength: 100
          example: New User
        githubId:
          type: string
          description: GitHub username (optional)
          nullable: true
          example: newuser
        avatar:
          type: string
          description: Avatar URL (optional)
          nullable: true
          example: https://example.com/avatar.jpg
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          description: Username (1-50 chars)
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z0-9_-]+$'
          example: updateduser
        displayName:
          type: string
          description: Display name (optional)
          nullable: true
          maxLength: 100
          example: Updated User
        githubId:
          type: string
          description: GitHub username (optional)
          nullable: true
          example: updateduser
        avatar:
          type: string
          description: Avatar URL (optional)
          nullable: true
          example: https://example.com/updated-avatar.jpg
...
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: User found
        data:
          $ref: '#/components/schemas/User'
        timestamp:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
    PaginatedUsersResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Users retrieved
        data:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: '#/components/schemas/User'
              example:
                - id: 1
                  username: john_doe
                  displayName: John Doe
                - id: 2
                  username: jane_smith
                  displayName: Jane Smith
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number (1-based)
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 10
        total:
          type: integer
          description: Total number of items
          example: 1234
        totalPages:
          type: integer
          description: Total number of pages
          example: 124
        hasNext:
          type: boolean
          description: Whether next page exists
          example: true
        hasPrev:
          type: boolean
          description: Whether previous page exists
          example: false
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        timestamp:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            type:
              type: string
              enum:
                [
                  BAD_REQUEST,
                  AUTHENTICATION_ERROR,
                  NOT_FOUND,
                  CONFLICT,
                  INTERNAL_ERROR,
                  SERVICE_UNAVAILABLE,
                  VALIDATION_ERROR,
                  ERROR,
                ]
              example: VALIDATION_ERROR
            message:
              type: string
              example: Username is required and must be a string
            timestamp:
              type: string
              format: date-time
              example: 2024-01-15T10:30:00Z
            requestId:
              type: string
              description: Request correlation ID
              example: req_1642245400000_a1b2c3d4
            errorId:
              type: string
              description: Unique error identifier
              example: err_a1b2c3d4
  responses:
    BadRequestError:
      description: Invalid request parameters or data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimitError:
      description: Too many requests - rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Maximum requests per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit window resets
          schema:
            type: integer
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  rateLimit:
                    type: object
                    properties:
                      limit:
                        type: integer
                        example: 100
                      remaining:
                        type: integer
                        example: 0
                      reset:
                        type: integer
                        example: 1642245460
                      window:
                        type: integer
                        example: 60
...
security:
  - BearerAuth:
    type: http
    scheme: bearer
    bearerFormat: JWT
    description: JWT token for authentication (if implemented)
  - ApiKeyAuth:
    type: apiKey
    in: header
    name: X-API-Key
    description: API key for service access (if implemented)